+++ reana_client/errors.py
@@ -15,3 +15,11 @@ class FileUploadError(Exception):
 
 class FileDeletionError(Exception):
     """File deletion didn't succeed."""
+
+
+class EnvironmentValidationError(Exception):
+    """REANA workflow environment validation didn't succeed."""
+
+
+class ParameterValidationError(Exception):
+    """REANA workflow parameter validation didn't succeed."""