+++ tests/test_validate_environments.py
@@ -10,7 +10,8 @@
 
 import pytest
 
-from reana_client.validation import _get_full_image_name, _validate_image_tag
+from reana_client.errors import EnvironmentValidationError
+from reana_client.validation.environments import SerialEnvironmentValidator
 
 
 @pytest.mark.parametrize(
@@ -26,15 +27,16 @@ from reana_client.validation import _get_full_image_name, _validate_image_tag
         ("foo:bar:", "has invalid tag", True),
     ],
 )
-def test_validate_environment_image_tag(image, output, exit_, capsys):
+def test_validate_environment_image_tag(image, output, exit_):
     """Validate workflow environment image tags."""
+    validator = SerialEnvironmentValidator()
     if exit_:
-        with pytest.raises(SystemExit):
-            _validate_image_tag(image)
+        with pytest.raises(EnvironmentValidationError) as e:
+            validator._validate_image_tag(image)
+        assert output in str(e)
     else:
-        _validate_image_tag(image)
-    captured = capsys.readouterr()
-    assert output in (captured.err if exit_ else captured.out)
+        validator._validate_image_tag(image)
+        assert output in validator.messages.pop()["message"]
 
 
 @pytest.mark.parametrize(
@@ -47,4 +49,5 @@ def test_validate_environment_image_tag(image, output, exit_, capsys):
     ],
 )
 def test_get_full_image_name(image, tag, full_image_name):
-    assert _get_full_image_name(image, tag) == full_image_name
+    validator = SerialEnvironmentValidator()
+    assert validator._get_full_image_name(image, tag) == full_image_name