+++ tests/test_validate_parameters.py
@@ -13,11 +13,9 @@ import yaml
 from click.testing import CliRunner
 
 from reana_client.utils import cwl_load
-from reana_client.validation import (
-    _validate_dangerous_operations,
-    _validate_serial_parameters,
-    _validate_yadage_parameters,
-    _validate_cwl_parameters,
+from reana_client.validation.parameters import (
+    SerialParameterValidator,
+    validate_parameters,
 )
 
 
@@ -30,6 +28,7 @@ def test_validate_parameters_cwl(
 ):
     """Validate parameters for CWL workflows."""
     runner = CliRunner()
+    workflow_type = "cwl"
 
     def get_loaded_yaml(step_spec, input_spec):
         with open("main.cwl", "w") as f:
@@ -46,9 +45,9 @@ def test_validate_parameters_cwl(
         reana_yaml = get_loaded_yaml(
             cwl_workflow_spec_step, cwl_workflow_spec_correct_input_param
         )
-        _validate_cwl_parameters(reana_yaml)
+        validate_parameters(workflow_type, reana_yaml)
         captured = capsys.readouterr()
-        assert not captured.out
+        assert "SUCCESS: Workflow operations appear valid." in captured.out
 
     # Wrong Input parameter used
     with runner.isolated_filesystem():
@@ -56,7 +55,7 @@ def test_validate_parameters_cwl(
             cwl_workflow_spec_step, cwl_workflow_spec_wrong_input_param
         )
         with pytest.raises(SystemExit) as exc_info:
-            _validate_cwl_parameters(reana_yaml)
+            validate_parameters(workflow_type, reana_yaml)
         assert (
             "Step is missing required parameter 'xoutputfile'" in exc_info.value.args[0]
         )
@@ -64,7 +63,7 @@ def test_validate_parameters_cwl(
     # Wrong file path used
     reana_yaml = yaml.load(create_cwl_yaml_workflow_schema, Loader=yaml.FullLoader)
     with pytest.raises(SystemExit) as exc_info:
-        _validate_cwl_parameters(reana_yaml)
+        validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert "ERROR: Workflow path main.cwl is not valid." in captured.err
 
@@ -72,13 +71,14 @@ def test_validate_parameters_cwl(
 def test_validate_parameters_serial(create_yaml_workflow_schema, capsys):
     """Validate parameters for Serial workflows."""
     reana_yaml = yaml.load(create_yaml_workflow_schema, Loader=yaml.FullLoader)
-    _validate_serial_parameters(reana_yaml)
+    workflow_type = "serial"
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
-    assert not captured.out
+    assert "Workflow parameters and commands appear valid." in captured.out
 
     # Input parameter not being used
     reana_yaml["inputs"]["parameters"]["foo"] = "foo"
-    _validate_serial_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: REANA input parameter "foo" does not seem to be used' in captured.out
@@ -89,15 +89,15 @@ def test_validate_parameters_serial(create_yaml_workflow_schema, capsys):
     reana_yaml["workflow"]["specification"]["steps"].append(
         {"commands": [r"$\{SHELL\} -c echo foo"]}
     )
-    _validate_serial_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
-    assert not captured.out
+    assert "Workflow parameters and commands appear valid." in captured.out
 
     # Parameter in unnamed step not defined
     reana_yaml["workflow"]["specification"]["steps"].append(
         {"commands": ["python ${foo} --bar"]}
     )
-    _validate_serial_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Serial parameter "foo" found on step "2" is not defined in input parameters'
@@ -108,7 +108,7 @@ def test_validate_parameters_serial(create_yaml_workflow_schema, capsys):
     reana_yaml["workflow"]["specification"]["steps"].append(
         {"commands": ["python ${bar} --foo"], "name": "baz"}
     )
-    _validate_serial_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Serial parameter "bar" found on step "baz" is not defined in input parameters'
@@ -119,7 +119,7 @@ def test_validate_parameters_serial(create_yaml_workflow_schema, capsys):
     reana_yaml["workflow"]["specification"]["steps"].append(
         {"commands": ["python ${bar} --foo"], "name": "qux"}
     )
-    _validate_serial_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert 'WARNING: Serial parameter "bar" found on steps' in captured.out
     assert "baz, qux" in captured.out or "qux, baz" in captured.out
@@ -129,13 +129,14 @@ def test_validate_parameters_serial(create_yaml_workflow_schema, capsys):
 def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
     """Validate parameters for Yadage workflows."""
     reana_yaml = yadage_workflow_spec_loaded
-    _validate_yadage_parameters(reana_yaml)
+    workflow_type = "yadage"
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
-    assert not captured.out
+    assert "Workflow parameters and commands appear valid." in captured.out
 
     # REANA input parameter not being used.
     reana_yaml["inputs"]["parameters"]["qux"] = "qux_val"
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: REANA input parameter "qux" does not seem to be used' in captured.out
@@ -155,7 +156,7 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
             },
         }
     )
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Yadage input parameter "qux" found on step "gendata" does not seem to be used.'
@@ -171,7 +172,7 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
     ]
     process["script"] += " && ./run-job {my_job}"
 
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Yadage parameter "my_job" found on step "gendata" is not defined in input parameters'
@@ -184,7 +185,7 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
     ]
     process["script"] += " && ./run-job {my_job}"
 
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert 'WARNING: Yadage parameter "my_job" found on steps' in captured.out
     assert "gendata, fitdata" in captured.out or "fitdata, gendata" in captured.out
@@ -195,7 +196,7 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
     subworkflow["workflow"]["stages"][0]["scheduler"]["parameters"].append(
         {"key": "subfoo", "value": "subfoo_val"}
     )
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Yadage input parameter "subfoo" found on step "nested_step" does not seem to be used.'
@@ -205,14 +206,14 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
     # Use previous parameter in command
     process = subworkflow["workflow"]["stages"][0]["scheduler"]["step"]["process"]
     process["script"] += " && go run {subfoo}"
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert "subfoo" not in captured.out
 
     # Parameter not defined in sub-step
     process = subworkflow["workflow"]["stages"][0]["scheduler"]["step"]["process"]
     process["script"] += " && go run {subbar}"
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Yadage parameter "subbar" found on step "nested_step" is not defined in input parameters'
@@ -224,7 +225,7 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
         "process"
     ]
     process["script"] += " && ./run-job {foo}"
-    _validate_yadage_parameters(reana_yaml)
+    validate_parameters(workflow_type, reana_yaml)
     captured = capsys.readouterr()
     assert (
         'WARNING: Yadage parameter "foo" found on step "fitdata" is not defined in input parameters.'
@@ -233,29 +234,30 @@ def test_validate_parameters_yadage(yadage_workflow_spec_loaded, capsys):
 
 
 @pytest.mark.parametrize(
-    "command, step, warning",
+    "commands, step, warning",
     [
-        ("python foo.py", "gendata", ""),
+        (["python foo.py"], "gendata", ""),
         (
-            "sudo python foo.py",
+            ["python foo.py", "sudo python foo.py"],
             "fitdata",
             '"sudo" found in step "fitdata" might be dangerous.',
         ),
         (
-            'echo "hello world!" && sudo python foo.py',
+            ['echo "hello world!" && sudo python foo.py'],
             "fitdata",
             '"sudo" found in step "fitdata" might be dangerous.',
         ),
         (
-            "cd /foo && npm install",
+            ["cd /foo && npm install", 'echo "hello world!"'],
             "installation",
             '"cd /" found in step "installation" might be dangerous.',
         ),
-        ("sudo npm install", None, '"sudo" might be dangerous.',),
+        (["sudo npm install"], None, '"sudo" might be dangerous.',),
     ],
 )
-def test_validate_dangerous_operations(command, step, warning, capsys):
+def test_validate_dangerous_operations(commands, step, warning, capsys):
     """Validate if dangerous operations in a command trigger a warning."""
-    _validate_dangerous_operations(command, step=step)
-    captured = capsys.readouterr()
-    assert warning in captured.out
+    validator = SerialParameterValidator({})
+    validator._validate_dangerous_operations(commands, step=step)
+    warnings = validator.operations_warnings
+    assert warning in (warnings.pop()["message"] if warnings else "")